refactor: 전시회를 위한 API 패키지 분리
